Refactored different docker containers into single container
