+++ backend/crud/main.py
@@ -3,6 +3,7 @@ import mysql.connector
 from flask_cors import CORS
 import datetime
 import requests
+import uuid
 
 app = Flask(__name__)
 CORS(app)
@@ -328,6 +329,98 @@ def add_disposition():
             cursor.close()
         if 'conn' in locals():
             conn.close()       
+
+@app.route("/search", methods=["GET"])
+def search_agents():
+    try:
+        # Get lat/lon from query parameters
+        lat = float(request.args.get('latitude'))
+        lon = float(request.args.get('longitude'))
+        booking_date = request.args.get('booking_date')
+        booking_time = request.args.get('booking_time')
+        booking_period = '02:00:00'
+        
+        conn = mysql.connector.connect(**db_config)
+        cursor = conn.cursor(dictionary=True, buffered=True)
+
+        # Build query based on whether postal_code is provided
+        if lat and lon:           
+            query = """
+                SELECT fa.name, fa.agentId,
+                    (6371 * ACOS(
+                        COS(RADIANS(%s)) * COS(RADIANS(l.latitude)) *
+                        COS(RADIANS(l.longitude) - RADIANS(%s)) +
+                        SIN(RADIANS(%s)) * SIN(RADIANS(l.latitude))
+                    )) AS distance
+                FROM field_agents fa
+                INNER JOIN locations l ON fa.location_id = l.id
+                WHERE fa.agentId NOT IN (
+                    SELECT b.agentId
+                    FROM bookings b
+                    WHERE b.booking_date = %s
+                      AND b.booking_time BETWEEN SUBTIME(%s, %s) AND ADDTIME(%s, %s)
+                )
+                ORDER BY distance ASC;
+            """
+
+            cursor.execute(query, (lat, lon, lat, booking_date, booking_time, booking_period, booking_time, booking_period))
+        else:
+            # Return all agents if no postal code specified
+            cursor.execute("SELECT * FROM field_agents")
+        
+        agents = cursor.fetchall()
+        return jsonify(agents), 200
+
+    except Exception as e:
+        return jsonify({"error": str(e)}), 500
+    finally:
+        if 'cursor' in locals():
+            cursor.close()
+        if 'conn' in locals():
+            conn.close()
+
+@app.route("/login", methods=["POST"])
+def login():
+    try:
+        data = request.get_json()
+        email = data.get("email")
+        password = data.get("password")
+
+        conn = mysql.connector.connect(**db_config)
+        cursor = conn.cursor(dictionary=True)
+
+        query = """
+        (
+            SELECT dispatcherId AS id, 'dispatcher' AS user_type
+            FROM dispatchers
+            WHERE email = %s AND password = %s
+        )
+        UNION
+        (
+            SELECT agentId AS id, 'agent' AS user_type
+            FROM field_agents
+            WHERE email = %s AND password = %s
+        )
+        """
+
+        cursor.execute(query, (email, password, email, password))
+        result = cursor.fetchone()
+
+        if result:
+            return {
+                "Login": "Successful",
+                "id": result["id"],
+                "user_type": result["user_type"],
+                "token": str(uuid.uuid4())
+            }, 200
+        else:
+            return {"error": "Email or password incorrect"}, 404
+
+    finally:
+        if 'cursor' in locals():
+            cursor.close()
+        if 'conn' in locals():
+            conn.close()
 
 def notify_customer(data):
     try:
