+++ backend/crud/requirements.txt
@@ -2,3 +2,4 @@ flask
 mysql-connector-python
 flask_cors
 requests
+uuid
