--- backend/login/Dockerfile
@@ -1,16 +1,0 @@
-FROM python:latest
-
-# Set the working directory
-WORKDIR /app
-
-# Copy the current directory contents into the container
-COPY requirements.txt .
-COPY wait-for-it.sh .
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-# Expose port for the Flask app
-EXPOSE 9000
-
-# Run the Flask app
-CMD ["./wait-for-it.sh", "mysql:3306", "--", "flask", "--debug", "run", "--host=0.0.0.0", "--port=9000"]
