--- backend/login/requirements.txt
@@ -1,4 +1,0 @@
-flask
-mysql-connector-python
-flask_cors
-uuid
