+++ frontend/src/api/login.ts
@@ -7,7 +7,7 @@ export interface LoginResponse {
 }
 
 export async function login(email: string, password: string): Promise<LoginResponse> {
-  const res = await fetch(`${BASE_URL}:9000/login`, {
+  const res = await fetch(`${BASE_URL}:8000/login`, {
     method: "POST",
     headers: { "Content-Type": "application/json" },
     body: JSON.stringify({ email, password}),
